+++ src/components/Header/ProfilePanel.tsx
@@ -2,8 +2,8 @@ import LogoutIcon from '@mui/icons-material/Logout';
 import NotificationsIcon from '@mui/icons-material/Notifications';
 import { Badge, Box, IconButton, Typography } from '@mui/material';
 import Avatar from '@mui/material/Avatar';
-import { useAuth } from 'hooks';
 import { AccountMenu } from 'components/Header/AccountMenu';
+import { useAuth } from 'providers/AuthProvider';
 
 export const ProfilePanel = () => {
   const { auth } = useAuth();
