+++ src/utils/getErrorMessage.ts
@@ -1,0 +1,4 @@
+export const getErrorMessage = (error: unknown) => {
+  if (error instanceof Error) return error.message;
+  return String(error);
+};
